+++ src/date-utils.ts
@@ -1,4 +1,4 @@
-import { DateTime } from "luxon";
+import { DateTime, IANAZone, LocalZone } from "luxon";
 import {Event} from "./event";
 import {RRule, rrulestr} from "rrule";
 
@@ -134,9 +134,9 @@ export function getDaysBetween(fstDate: Date, sndDate: Date): Number {
  * @param originalEvent - The original event.
  * @param rrule - The RRULE (coming from ICS) for the original event.
  */
-export function getRecurringEvents(originalEvent: Event, rrule: string): Event[] {
+export function getRecurringEvents(originalEvent: Event, rrule: any): Event[] {
   const today: Date = new Date();
-  let rule = rrulestr(`RRULE:${rrule}`);
+  let rule = rrule; //rrulestr(`RRULE:${rrule}`);
   const origOptions = rule.origOptions;
   const originalStartDate: Date = new Date(originalEvent.startDate);
   origOptions.dtstart = originalStartDate;
@@ -214,4 +214,35 @@ export function removeChangedEventsFromRecurringEvents(events: Event[]) {
       delete events[i].originalUID;
     }
   }
+}
+
+function getRecurringStartDates(event: any) {
+  const result: Date[] = [];
+  const dates = event.rrule.between(new Date(2021, 0, 1, 0, 0, 0, 0), new Date(2042, 11, 31, 0, 0, 0, 0))
+  if (dates.length === 0) return result;
+
+  console.log('Summary:', event.summary);
+  console.log('Original start:', event.start);
+  console.log('RRule start:', `${event.rrule.origOptions.dtstart} [${event.rrule.origOptions.tzid}]`)
+
+  dates.forEach((date: Date) => {
+    let newDate
+    if (event.rrule.origOptions.tzid) {
+      // tzid present (calculate offset from recurrence start)
+      const dateTimezone = IANAZone.create('UTC');// moment.tz.zone('UTC')
+      const localTimezone = new LocalZone(); //moment.tz.guess()
+      const tz = localTimezone // event.rrule.origOptions.tzid === localTimezone ? event.rrule.origOptions.tzid : localTimezone // TODO
+      const timezone = localTimezone;
+      const offset = timezone.offset(date.getTime()) - dateTimezone.offset(date.getTime()); // timezone.utcOffset(date) - dateTimezone.utcOffset(date)
+      newDate = DateTime.fromJSDate(date).plus({minutes: offset}).toJSDate(); //moment(date).add(offset, 'minutes').toDate()
+    } else {
+      // tzid not present (calculate offset from original start)
+      newDate = new Date(date.setHours(date.getHours() - ((event.start.getTimezoneOffset() - date.getTimezoneOffset()) / 60)))
+    }
+
+    console.log('Recurrence start:', newDate);
+    result.push(newDate);
+  });
+
+  return result;
 }