+++ src/ics-to-json-converter.ts
@@ -11,7 +11,7 @@ const md5 = require('md5');
 import { Event } from './event';
 import {getRecurringEvents, removeChangedEventsFromRecurringEvents} from "./date-utils";
 
-const ICAL = require("ical.js");
+const ical = require("node-ical");
 const outputType = "application/json";
 
 /**
@@ -19,12 +19,8 @@ const outputType = "application/json";
  */
 export class IcsToJsonConverter extends BaseTypedRepresentationConverter {
 
-  private removeAppleLocation: boolean;
-  public constructor(
-    options?: { removeAppleLocation?: boolean }
-  ) {
+  public constructor() {
     super("text/calendar", outputType);
-    this.removeAppleLocation = options?.removeAppleLocation || false;
   }
 
   public async handle({
@@ -37,29 +33,24 @@ export class IcsToJsonConverter extends BaseTypedRepresentationConverter {
     if (!data?.length)
       throw new BadRequestHttpError("Empty input is not allowed");
 
-    if (this.removeAppleLocation) {
-      data = data.replace(/X-APPLE-STRUCTURED-LOCATION;(((?!END:VEVENT).|\r|\n)*)/g, '');
-      //console.log(data);
-    }
-
-    const jcalData = ICAL.parse(data);
-    const vcalendar = new ICAL.Component(jcalData);
-    const vevents = vcalendar.getAllSubcomponents("vevent");
+    const parsedICS = await ical.async.parseICS(data);
+    const vcalendar = parsedICS.vcalendar;
+    const vevents = this.getVEvents(parsedICS);
 
     for (const vevent of vevents) {
-      const summary = vevent.getFirstPropertyValue("summary");
+      const {summary} = vevent;
 
       if (!summary) {
-        console.warn(`Encountered event without summary/title in calendar "${vcalendar.getFirstPropertyValue("x-wr-calname")}".`);
+        console.warn(`Encountered event without summary/title in calendar "${vcalendar['WR-CALNAME']}".`);
       }
 
-      let startDate = vevent.getFirstPropertyValue("dtstart");
+      let startDate = vevent.start;
 
       if (!startDate)
         throw new BadRequestHttpError("Dtstart needs to be provided");
 
       startDate = new Date(startDate);
-      let endDate = vevent.getFirstPropertyValue("dtend");
+      let endDate = vevent.end;
       endDate = new Date(endDate);
 
       const event: Event = {
@@ -69,29 +60,28 @@ export class IcsToJsonConverter extends BaseTypedRepresentationConverter {
         hash: md5(summary + startDate + endDate)
       };
 
-      if (vevent.hasProperty("description"))
-        event.description = vevent.getFirstPropertyValue("description");
-      if (vevent.hasProperty("url"))
-        event.url = vevent.getFirstPropertyValue("url");
-      if (vevent.hasProperty("location"))
-        event.location = vevent.getFirstPropertyValue("location");
-      if (vevent.hasProperty("uid"))
-        event.originalUID = vevent.getFirstPropertyValue("uid");
-      if (vevent.hasProperty("recurrence-id"))
-        event.originalRecurrenceID = vevent.getFirstPropertyValue("recurrence-id");
+      event.description = vevent.description || undefined;
+      event.url = vevent?.url?.val || vevent?.url || undefined;
+      event.location = vevent.location || undefined;
+      event.originalUID = vevent.uid || undefined;
+      event.originalRecurrenceID = vevent.recurrenceid || undefined;
 
       events.push(event);
 
-      if (vevent.getFirstPropertyValue("rrule")) {
-        const recurringEvents = getRecurringEvents(event, vevent.getFirstPropertyValue("rrule"));
+      if (vevent.rrule) {
+        const recurringEvents = getRecurringEvents(event, vevent.rrule);
         events = events.concat(recurringEvents);
       }
     }
 
     removeChangedEventsFromRecurringEvents(events);
 
+    if (!vcalendar) {
+      throw new InternalServerError("No calendar found");
+    }
+
     const calendar = {
-      name: vcalendar.getFirstPropertyValue("x-wr-calname") as string,
+      name: vcalendar["WR-CALNAME"] as string,
       events,
     };
 
@@ -103,5 +93,29 @@ export class IcsToJsonConverter extends BaseTypedRepresentationConverter {
       representation.metadata,
       outputType
     );
+  }
+
+  private getVEvents(data: any) {
+    const events = [];
+
+    for (const id in data) {
+      const el = data[id];
+
+      if (el.type === 'VEVENT') {
+        events.push(el);
+
+        if (el.recurrences) {
+          for (const id2 in el.recurrences) {
+            const el2 = el.recurrences[id2];
+
+            if (el2.type === 'VEVENT') {
+              events.push(el2);
+            }
+          }
+        }
+      }
+    }
+
+    return events;
   }
 }